Clean things up a bit and update CHANGELOG.md
